tests/wpt/css-tests/css-transforms-1_dev/xhtml1/transform_translate_max.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <title>CSS Transforms API Test: transform translate with maximum translation value</title>
    <link href="http://www.intel.com" rel="author" title="Intel" />
    <link href="http://www.w3.org/TR/css-transforms-1/#transform-property" rel="help" />
    <link href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions" rel="help" />
    <meta content="dom" name="flags" />
    <meta content="Check if transform sets translate(INFINITE, INFINITE),
                                 transform property returns initial value." name="assert" />
    <script src="/resources/testharness.js" type="text/javascript"></script>
    <script src="/resources/testharnessreport.js" type="text/javascript"></script>
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
    test(function() {
        document.getElementById("test").style.transform = "translate(INFINITE, INFINITE)";
        var value = window.getComputedStyle(document.getElementById("test")).getPropertyValue("transform");
        assert_equals(value, "none", "transform property value should be none");
    }, "transform_translate_max");
    </script>


</body></html>
